+++ package.json
@@ -8,10 +8,13 @@
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "express": "^4.17.1",
+    "node-sass": "^5.0.0",
+    "normalize-scss": "^7.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "sass": "^7.6.3",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
@@ -19,7 +22,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "cypress": "npx open cypress"
+    "cypress": "npx open cypress",
+    "sass": "sass src/Sass:src/Css --watch --no-source-map"
   },
   "eslintConfig": {
     "extends": [
@@ -41,6 +45,7 @@
   },
   "devDependencies": {
     "cypress": "^6.8.0",
-    "nodemon": "^2.0.7"
+    "nodemon": "^2.0.7",
+    "sass": "^1.32.8"
   }
 }
